+++ unlock-app/src/components/interface/locks/List/elements/LockList.tsx
@@ -119,7 +119,7 @@ export const LockList = ({ owner }: LockListProps) => {
     ]
   }, [networks, defaultNetwork])
 
-  const results = useLocksByManagerOnNetworks(owner, networkItems)
+  const results = useLocksByManagerOnNetworks(owner, networkItems, 'list')
   const result = results[0]
 
   const [favoriteLocks, setFavoriteLocks] = useState<FavoriteLocks>(() => {
