+++ unlock-app/src/hooks/useCentralizedLockData.ts
@@ -15,7 +15,7 @@ export const useCentralizedLockData = (
   }
 ) => {
   return useQuery({
-    queryKey: ['centralizedLockData', lockAddress, network],
+    queryKey: ['centralizedLockData', lockAddress, network, 'management'], // Added context to avoid collisions
     // Only run this query when we have valid lock parameters
     enabled: !!lockAddress && !!network && !!owner,
     queryFn: async () => {
@@ -62,8 +62,11 @@ export const useCentralizedLockData = (
         metadata: metadataResponse?.data || {},
       }
     },
-    staleTime: options.staleTime, // Configurable stale time with reasonable default
+    staleTime: options.staleTime || 10 * 60 * 1000, // Use provided staleTime or default to 10 minutes
+    gcTime: 30 * 60 * 1000, // 30 minutes cache time
     refetchOnWindowFocus: false,
+    refetchOnMount: false, // Prevent refetching on mount
+    retry: false, // Prevent retries which can cause multiple fetches
   })
 }
