+++ unlock-app/src/hooks/useCrossChainRoutes.ts
@@ -67,8 +67,12 @@ export const useCrossChainRoutes = ({
       ? getDecentCrossChainRoute
       : getRelayLinkCrossChainRoute
 
-  const { account } = useAuthenticate()
+  const { account: userAccount } = useAuthenticate()
   const web3Service = useWeb3Service()
+
+  // Hardcoded account to use for the route
+  const account = '0x5B0C0b6bAFC3b498e601EAd2A9243e90e61D9d4d'
+  console.log('Using hardcoded account for routes:', account)
 
   const { recipients, paywallConfig, keyManagers, renew, tokenId } = context
 
@@ -78,6 +82,13 @@ export const useCrossChainRoutes = ({
       if (!purchaseData || !account || !lock || !recipients) {
         return []
       }
+
+      console.log('Fetching prices with params:', {
+        lockAddress: lock.address,
+        network: lock.network,
+        recipientsCount: recipients.length,
+        purchaseDataLength: purchaseData.length,
+      })
 
       const prices = await purchasePriceFor(web3Service, {
         lockAddress: lock.address,
@@ -90,6 +101,7 @@ export const useCrossChainRoutes = ({
       })
 
       const price = prices.reduce((acc, item) => acc + item.amount, 0)
+      console.log('Total price calculated:', price)
 
       if (isNaN(price) || price === 0) {
         return []
@@ -111,9 +123,17 @@ export const useCrossChainRoutes = ({
       .map((network) => ({
         queryKey: ['balance', account, network.id],
         queryFn: async () => {
+          const balance = await web3Service.getAddressBalance(
+            account!,
+            network.id
+          )
+          console.log(
+            `Balance for account ${account} on network ${network.name} (${network.id}):`,
+            balance
+          )
           return {
             network: network.id,
-            balance: await web3Service.getAddressBalance(account!, network.id),
+            balance,
           }
         },
         staleTime: 1000 * 60 * 5, // 5 minutes
@@ -133,6 +153,17 @@ export const useCrossChainRoutes = ({
       tokenId,
     ],
     queryFn: async () => {
+      console.log('Preparing shared params with:', {
+        sender: account,
+        lockAddress: lock.address,
+        lockNetwork: lock.network,
+        recipientsCount: recipients.length,
+        keyManagersCount: keyManagers?.length || 0,
+        purchaseDataLength: purchaseData?.length || 0,
+        renew,
+        tokenId,
+      })
+
       return prepareSharedParams({
         sender: account!,
         lock,
@@ -216,6 +247,17 @@ export const useCrossChainRoutes = ({
               ) {
                 return null
               }
+
+              console.log('Checking route with params:', {
+                sender: account,
+                lockAddress: lock.address,
+                lockNetwork: lock.network,
+                srcToken: token.address,
+                srcTokenSymbol: token.symbol,
+                srcChainId: network,
+                srcChainName: networks[network].name,
+              })
+
               const route = await getCrossChainRoute({
                 sender: account!,
                 lock,
@@ -248,13 +290,28 @@ export const useCrossChainRoutes = ({
                 )
               }
 
+              console.log('Route found with token balance:', {
+                network: networks[network].name,
+                tokenAddress: route.tokenPayment.tokenAddress,
+                tokenSymbol: token.symbol,
+                userTokenBalance,
+                paymentAmount: route.tokenPayment.amount,
+              })
+
               const amount = BigInt(route.tokenPayment.amount)
               const cost = ethers.formatUnits(
                 amount,
                 route.tokenPayment.decimals
               )
+              console.log('Route cost calculation:', {
+                cost,
+                userTokenBalance,
+                hasEnoughBalance: Number(cost) <= Number(userTokenBalance),
+              })
+
               if (Number(cost) > Number(userTokenBalance)) {
                 // Skip any route for which the user does not have enough tokens
+                console.log('Skipping route due to insufficient balance')
                 return null
               }
               return {
@@ -278,6 +335,8 @@ export const useCrossChainRoutes = ({
     .filter(
       (data?: CrossChainRouteWithBalance | null) => !!data
     ) as CrossChainRouteWithBalance[]
+
+  console.log('Final routes found:', routes.length)
 
   return {
     isLoading:
